Refactor schema validation into a modular package structure with backward compatibility
