+++ SCHEMA_VALIDATOR_REFACTORING.md
@@ -1,0 +1,78 @@
+# Schema Validator Refactoring Summary
+
+## Overview
+Successfully refactored the large 677-line `schema_validator.py` file into a well-organized modular package structure.
+
+## New Package Structure
+
+```
+src/forklift/processors/schema_validator/
+├── __init__.py                 # Package exports and backward compatibility
+├── base_local.py              # Local base classes (ValidationResult, BaseProcessor)
+├── config.py                  # Configuration classes and enums
+├── constraints.py             # Constraint validation logic
+├── core.py                    # Main SchemaValidator class
+├── schema.py                  # Schema definition classes
+├── type_converter.py          # Type conversion utilities
+└── utils.py                   # Utility functions
+```
+
+## Modular Components
+
+### 1. Configuration (`config.py`)
+- `SchemaValidationMode` enum (STRICT, PERMISSIVE, COERCE)
+- `NullabilityMode` enum (ERROR, WARNING, IGNORE)
+- `SchemaValidatorConfig` dataclass with all validation settings
+
+### 2. Schema Definitions (`schema.py`)
+- `ColumnSchema` dataclass for individual column specifications
+
+### 3. Type Conversion (`type_converter.py`)
+- `TypeConverter` class with static methods for:
+  - Converting between string types and PyArrow types
+  - Arrow schema to/from dictionary conversions
+  - Type compatibility checking
+  - Type coercion validation
+
+### 4. Constraint Validation (`constraints.py`)
+- `ConstraintValidator` class with static methods for:
+  - Range constraints (min/max values)
+  - Enum constraints (allowed values)
+  - Pattern constraints (regex validation)
+  - Length constraints (string length validation)
+
+### 5. Core Validation (`core.py`)
+- Main `SchemaValidator` class with all validation logic:
+  - Batch structure validation
+  - Column presence validation
+  - Data type validation
+  - Nullability validation
+  - Constraint validation
+  - Row count validation
+
+### 6. Utilities (`utils.py`)
+- `create_schema_validator_from_json()` - Create validator from JSON schema
+- `create_schema_from_batch()` - Generate schema from PyArrow batch
+
+### 7. Backward Compatibility
+- Original `schema_validator.py` now imports from the new modular structure
+- All existing code using the schema validator will continue to work unchanged
+- Maintains the same public API
+
+## Benefits of Refactoring
+
+1. **Modularity**: Each concern is separated into its own focused module
+2. **Maintainability**: Easier to find, understand, and modify specific functionality
+3. **Testability**: Individual components can be tested in isolation
+4. **Reusability**: Components can be imported and used independently
+5. **Scalability**: New validation features can be added without affecting existing code
+6. **Code Organization**: Clear separation of configuration, validation logic, and utilities
+
+## Testing Results
+- ✅ All imports work correctly
+- ✅ Backward compatibility maintained
+- ✅ Core functionality validated with test data
+- ✅ No circular import issues
+- ✅ Clean package structure
+
+The refactoring transforms a monolithic 677-line file into 8 focused, well-organized modules while maintaining complete backward compatibility.
